+++ backend/src/message/message.module.ts
@@ -6,9 +6,11 @@ import { ConversationModel } from "./models/conversation.model";
 import { MessageModel } from "./models/message.model";
 import { UserConversationModel } from "./models/user-conversation.model";
 import { ActiveConversationModel } from "./models/active-conversation";
+import { AuthModule } from "../auth/auth.module";
 
 @Module({
   imports: [
+    AuthModule,
     SequelizeModule.forFeature([
       ConversationModel,
       MessageModel,
@@ -17,5 +19,6 @@ import { ActiveConversationModel } from "./models/active-conversation";
     ]),
   ],
   providers: [MessageGateway, ConversationService],
+  exports: [ConversationService],
 })
 export class MessageModule {}