+++ backend/src/role/models/role.model.ts
@@ -9,7 +9,6 @@ interface RoleCreationAttrs {
 @Table({ tableName: "roles", createdAt: false, updatedAt: false })
 export class RoleModel extends Model<RoleModel, RoleCreationAttrs> {
   @ApiProperty({ example: "5", description: "Unique identifier" })
-  @Column({ type: DataType.INTEGER })
   @Column({
     type: DataType.INTEGER,
     unique: true,
@@ -19,12 +18,10 @@ export class RoleModel extends Model<RoleModel, RoleCreationAttrs> {
   id: number;
 
   @ApiProperty({ example: "5", description: "Role's value" })
-  @Column({ type: DataType.INTEGER })
   @Column({ type: DataType.STRING, unique: true, allowNull: false })
   value: string;
 
   @ApiProperty({ example: "5", description: "Role's description" })
-  @Column({ type: DataType.INTEGER })
   @Column({ type: DataType.STRING, allowNull: false })
   description: string;
 }